refactor: App fsd 적용
