+++ src/index.tsx
@@ -1,7 +1,8 @@
 import React from "react"
 import ReactDOM from "react-dom/client"
 import { BrowserRouter as Router } from "react-router-dom"
-import App from "./App"
+
+import { App } from "@/app"
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
